* Internal ctors for sinks with injectable dependencies for tests * Extracted DataTableCreator from MSSqlServerSinkTraits
