+++ src/Serilog.Sinks.MSSqlServer/Sinks/MSSqlServer/Platform/IDataTableCreator.cs
@@ -1,0 +1,9 @@
+﻿using System.Data;
+
+namespace Serilog.Sinks.MSSqlServer.Platform
+{
+    internal interface IDataTableCreator
+    {
+        DataTable CreateDataTable(string tableName, ColumnOptions columnOptions);
+    }
+}
